% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl_ops.R, R/scratch.R
\name{ancestors_names}
\alias{ancestors_names}
\title{obtain list of names of a set of ancestors}
\usage{
ancestors_names(anclist)

ancestors_names(anclist)
}
\arguments{
\item{anclist}{output of `ancestors`}
}
\value{
list of vectors of character()

list of vectors of character()
}
\description{
obtain list of names of a set of ancestors

obtain list of names of a set of ancestors
}
\note{
non-entities are removed and names are extracted

non-entities are removed and names are extracted
}
\examples{
pa = get_ordo_owl_path()
orde = setup_entities(pa)
al = ancestors(orde[1001:1002])
ancestors_names(al)
pa = get_ordo_owl_path()
orde = setup_entities(pa)
al = ancestors(orde[1001:1002])
ancestors_names(al)
}
