% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scratch.R
\name{ancestors_work}
\alias{ancestors_work}
\title{construct owlents instance from an owl file and then retrieve ancestor 'sets'}
\usage{
ancestors_work(owlfn)
}
\arguments{
\item{owlfn}{character(1) path to valid owl ontology}
}
\value{
a list of sets
}
\description{
construct owlents instance from an owl file and then retrieve ancestor 'sets'
}
\examples{
pa = get_ordo_owl_path()
ancestors(pa)
}
