% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_ontos.R
\name{available_ontos}
\alias{available_ontos}
\title{interrogate the cache for owl files and serialized ontologyIndex instances}
\usage{
available_ontos(ca = BiocFileCache::BiocFileCache(), token = "owl|OIRDS")
}
\arguments{
\item{ca}{defaults to BiocFileCache::BiocFileCache()}

\item{token}{character(1) defaults to "owl|OIRDS"}
}
\value{
NULL if nothing relevant is found in cache, otherwise a data.frame
with all cached information
}
\description{
interrogate the cache for owl files and serialized ontologyIndex instances
}
