% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioregistry.R
\name{bioregistry_ols_resources}
\alias{bioregistry_ols_resources}
\title{produce bioregistry_ols table}
\usage{
bioregistry_ols_resources()
}
\value{
data.frame
}
\description{
produce bioregistry_ols table
}
\note{
This uses the `resources` method of the bioregistry module from pip
to isolate resources with a non-null `ols` component.
}
\examples{
tab = bioregistry_ols_resources()
head(tab[,1:3])
}
