% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_ontos.R
\name{browse_ontos}
\alias{browse_ontos}
\title{browse available ontologies with datatable}
\usage{
browse_ontos(ca = BiocFileCache::BiocFileCache(), token = "owl|OIRDS")
}
\arguments{
\item{ca}{defaults to BiocFileCache::BiocFileCache()}

\item{token}{character(1) defaults to "owl|OIRDS"}
}
\description{
browse available ontologies with datatable
}
