% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl_ops.R
\name{children_names}
\alias{children_names}
\title{obtain list of names of a set of subclasses/children}
\usage{
children_names(sclist)
}
\arguments{
\item{sclist}{output of `subclasses`}
}
\value{
list of vectors of character()
}
\description{
obtain list of names of a set of subclasses/children
}
\note{
non-entities are removed and names are extracted
}
\examples{
pa = get_ordo_owl_path()
o2 = try(reticulate::import("owlready2"), silent=TRUE)
if (!inherits(o2, "try-error")) {
 orde = setup_entities(pa)
 al = subclasses(orde[100:120])
 children_names(al)
}
}
