% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{get_classlabs}
\alias{get_classlabs}
\title{return a vector of class labels in an ontology}
\usage{
get_classlabs(owlfile)
}
\arguments{
\item{owlfile}{reference to OWL file, can be URL, will be processed
by owlready2.get_ontology}
}
\value{
vector
}
\description{
return a vector of class labels in an ontology
}
\examples{
pa = get_ordo_owl_path()
labs = get_classlabs(pa)
head(labs)
}
