% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{get_classnames}
\alias{get_classnames}
\title{return a vector of class names in an ontology}
\usage{
get_classnames(owlfile)
}
\arguments{
\item{owlfile}{reference to OWL file, can be URL, will be processed
by owlready2.get_ontology}
}
\value{
vector
}
\description{
return a vector of class names in an ontology
}
\examples{
pa = get_ordo_owl_path()
cls = get_classnames(pa)
head(cls)
}
