% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ordo_owl_path.R
\name{get_ordo_owl_path}
\alias{get_ordo_owl_path}
\title{decompress ordo owl file}
\usage{
get_ordo_owl_path(target = tempdir())
}
\arguments{
\item{target}{character(1) path to where decompressed owl will live}
}
\description{
decompress ordo owl file
}
