% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAvailOntos.R
\name{getAvailOntos}
\alias{getAvailOntos}
\title{This function returns a list of available ontologies in the ontoProc package.}
\usage{
getAvailOntos()
}
\description{
This function returns a list of available ontologies in the ontoProc package.
}
\examples{
getAvailOntos()
}
