% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntos.R
\name{getChebiLite}
\alias{getChebiLite}
\alias{getCellosaurusOnto}
\alias{getUBERON_NE}
\alias{getChebiOnto}
\alias{getOncotreeOnto}
\alias{getDiseaseOnto}
\alias{getGeneOnto}
\alias{getHCAOnto}
\alias{getPROnto}
\alias{getPATOnto}
\alias{getMondoOnto}
\alias{getSIOOnto}
\title{basic getters in old style, retained 2023 for deprecation interval}
\usage{
getChebiLite()

getCellosaurusOnto()

getUBERON_NE()

getChebiOnto()

getOncotreeOnto()

getDiseaseOnto()

getGeneOnto()

getHCAOnto()

getPROnto()

getPATOnto()

getMondoOnto()

getSIOOnto()
}
\value{
instance of ontology_index (S3) from ontologyIndex
}
\description{
basic getters in old style, retained 2023 for deprecation interval
}
\note{
getChebiOnto loads ontoRda/chebi_full.rda

getOncotreeOnto loads ontoRda/oncotree.rda

getDiseaseOnto loads ontoRda/diseaseOnto.rda

getHCAOnto loads ontoRda/hcaOnto.rda produced from hcao.owl at https://github.com/HumanCellAtlas/ontology/releases/tag/1.0.6 2/11/2019,
python pronto was used to convert OWL to OBO.

getPROnto loads ontoRda/PRonto.rda, produced from http://purl.obolibrary.org/obo/pr.obo 'reasoned' ontology from OBO foundry, 02-08-2019.
In contrast to other ontologies, this is imported via get_OBO with
`extract_tags='minimal'`.

getPATOnto loads ontoRda/patoOnto.rda, produced from https://raw.githubusercontent.com/pato-ontology/pato/master/pato.obo from OBO foundry, 02-08-2019.
}
