% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntos.R
\name{getOnto}
\alias{getOnto}
\title{get the ontology based on a short tag and year}
\usage{
getOnto(ontoname = "cellOnto", year_added = "2022")
}
\arguments{
\item{ontoname}{character(1) must be an element in `valid_ontonames()`}

\item{year_added}{character(1) refers to `rdatadateadded` in AnnotationHub metadata}
}
\description{
get the ontology based on a short tag and year
}
\note{
This queries AnnotationHub for "ontoProcData" and then filters to find
the AnnotationHub accession number and retrieves the ontologyIndex serialization
of the associated OBO representation of the ontology.
}
\examples{
co = getOnto()
tail(co$name[1000:1500])
}
