% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphNEL.R
\name{onto_plot2}
\alias{onto_plot2}
\title{high-level use of graph/Rgraphviz for rendering ontology relations}
\usage{
onto_plot2(ont, terms2use, cex = 0.8, width = 15, ...)
}
\arguments{
\item{ont}{instance of ontology from ontologyIndex}

\item{terms2use}{character vector}

\item{cex}{numeric(1) defaults to .8, supplied to Rgraphviz::graph.par}

\item{width}{integer(1) maximum line width for wrapping term names,
passed to \code{\link{improveNodes}}, defaults to 15}

\item{...}{passed to onto_plot of ontologyPlot}
}
\value{
graphNEL instance (invisibly)
}
\description{
high-level use of graph/Rgraphviz for rendering ontology relations
}
\examples{
cl = getOnto("cellOnto")
cl3k = c("CL:0000492", "CL:0001054", "CL:0000236", "CL:0000625",
   "CL:0000576", "CL:0000623", "CL:0000451", "CL:0000556")
onto_plot2(cl, cl3k)
}
