% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\docType{data}
\name{ontoProc_env}
\alias{ontoProc_env}
\title{OntoProc_env utilize the basilisk package to}
\format{
An object of class \code{BasiliskEnvironment} of length 1.
}
\usage{
ontoProc_env
}
\description{
OntoProc_env utilize the basilisk package to
}
\keyword{datasets}
