% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{packDesc2021}
\alias{packDesc2021}
\title{packDesc2021: overview of ontoProc resources}
\format{
data.frame instance
}
\usage{
packDesc2021
}
\description{
packDesc2021: overview of ontoProc resources
}
\note{
Brief survey of functions available to load serialized
ontology_index instances imported from OBO.  Focus is on versions added in 2021.
}
\examples{
data(packDesc2021)
head(packDesc2021)
}
\keyword{datasets}
