% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{packDesc2022}
\alias{packDesc2022}
\title{packDesc2022: overview of ontoProc resources}
\format{
data.frame instance
}
\usage{
data(packDesc2022)
}
\description{
packDesc2022: overview of ontoProc resources
}
\note{
Brief survey of functions available to load serialized
ontology_index instances imported from OBO.  Focus is on versions added in 2022.
}
\examples{
data(packDesc2022)
head(packDesc2022)
}
\keyword{datasets}
