% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{packDesc2023}
\alias{packDesc2023}
\title{packDesc2023: overview of ontoProc resources}
\format{
data.frame instance
}
\usage{
data(packDesc2023)
}
\description{
packDesc2023: overview of ontoProc resources
}
\note{
Brief survey of functions available to load serialized
ontology_index instances imported from OBO.  Focus is on versions added in 2023.
Several manual interventions were needed -- cellosaurus was too large to use the
script in inst/scripts/desc.R, and a number of ontologies do not have 2023 versions.
}
\examples{
data(packDesc2023)
head(packDesc2023)
}
\keyword{datasets}
