% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl_ops.R
\name{parents}
\alias{parents}
\title{retrieve is_a}
\usage{
parents(oe)
}
\arguments{
\item{oe}{owlents instance}
}
\value{
list of vectors of tags of parents
}
\description{
retrieve is_a
}
\examples{
pa = get_ordo_owl_path()
orde = setup_entities(pa)
orde
parents(orde[1000:1001])
labels(orde[1000:1001])
}
