% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickOnto.R
\name{quickOnto}
\alias{quickOnto}
\title{Probe the cache for entries found by bfcquery using the given query.}
\usage{
quickOnto(query, cache = BiocFileCache::BiocFileCache(), qans.only = FALSE)
}
\arguments{
\item{query}{character(1), will be used by BiocFileCache::bfcquery}

\item{cache}{instance of class(BiocFileCache::BiocFileCache())}

\item{qans.only}{logical, defaulting to FALSE, if TRUE, just return
the bfcquery result.}
}
\description{
Probe the cache for entries found by bfcquery using the given query.
}
\note{
Take only the last entry found if there are multiple hits.  Use
`setup_entities2` to transform owl to ontologyIndex and cache the
result if this is absent for the owl identified by query.
}
\examples{
aeo = quickOnto("aeo.owl")
str(aeo)
}
