% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl_ops.R
\name{search_labels}
\alias{search_labels}
\title{use owlready2 ontology search facility on term labels}
\usage{
search_labels(oents, regexp, case_sensitive = TRUE)
}
\arguments{
\item{oents}{owlents instance}

\item{regexp}{character(1) simple regular expression}

\item{case_sensitive}{logical(1) should case be respected in search?}
}
\value{
A named list: term labels are elements, tags are names of elements.
Will return NULL if nothing is found.
}
\description{
use owlready2 ontology search facility on term labels
}
\examples{
pa = get_ordo_owl_path()
orde = setup_entities(pa)
ol = search_labels(orde, "*Immunog*")
plot(orde, names(ol))
}
