% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl_ops.R
\name{[.owlents}
\alias{[.owlents}
\title{subset method}
\usage{
\method{[}{owlents}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{owlents instance}

\item{i}{character or numeric vector}

\item{j}{not used}

\item{drop}{not used}
}
\description{
subset method
}
