% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl2cache.R
\name{url_ok}
\alias{url_ok}
\title{check that a URL can get a 200 for a HEAD request}
\usage{
url_ok(url)
}
\arguments{
\item{url}{character(1)}
}
\value{
logical(1)
}
\description{
check that a URL can get a 200 for a HEAD request
}
