% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntos.R
\name{valid_ontonames}
\alias{valid_ontonames}
\title{give a vector of valid 'names' of ontoProc ontologies}
\usage{
valid_ontonames()
}
\description{
give a vector of valid 'names' of ontoProc ontologies
}
\examples{
head(valid_ontonames())
}
