% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formalize.R
\name{formalize}
\alias{formalize}
\title{use an LLM to match informal terms to terms in an ontology}
\usage{
formalize(
  informal_terms,
  ontology_terms,
  ontology_tags,
  llm_model = "gpt-4.1-2025-04-14"
)
}
\arguments{
\item{informal_terms}{character() vector of terms not necessarily found in ontology}

\item{ontology_terms}{character() vector of ontology terms}

\item{ontology_tags}{character() vector of tags for ontology terms, must be of same length as ontology_terms}

\item{llm_model}{character(1) used with chat_openai in ellmer, defaults to "gpt-4.1-2025-04-14"}
}
\description{
use an LLM to match informal terms to terms in an ontology
}
\note{
expects to have OPENAI_API_KEY set
}
\examples{
if (interactive()) {
 ctypes = c("tPlasma cells", "tMoMacDC", "tT cells",   # from Zilionis
  "tB cells", "tNK cells", "tNeutrophils", "Fibroblasts", "Type II cells", 
  "tpDC", "Endothelial cells", "tMast cells", "Smooth muscle cells", 
  "ND", "Club cells", "bNeutrophils", "bT cells", "bMonocytes", 
  "bNK cells", "bRBC", "bpDC", "bB cells", "bPlasma cells", "bPlatelets", 
  "tRBC", "Type I cells", "Ciliated cells", "bBasophils")
 cc = owl2cache(url="http://purl.obolibrary.org/obo/cl.owl")
 cloi = setup_entities2(cc)
 oname = cloi$name
 actual = grep("CL_", names(oname))
 oterms = as.character(oname[actual])
 otags = names(oname[actual])
 octy = formalize(ctypes, oterms, otags)
 head(octy)
 onto_plot2(cloi, unique(na.omit(octy$tag)))
}
}
