% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl_ops.R
\name{search_labels}
\alias{search_labels}
\title{use owlready2 ontology search facility on term labels}
\usage{
search_labels(ontopath, regexp, case_sensitive = TRUE)
}
\arguments{
\item{ontopath}{character(1) path to owl file}

\item{regexp}{character(1) simple regular expression}

\item{case_sensitive}{logical(1) should case be respected in search?}
}
\value{
A named list: term labels are elements, tags are names of elements.
Will return NULL if nothing is found.
}
\description{
use owlready2 ontology search facility on term labels
}
\examples{
pa = get_ordo_owl_path()
ol = search_labels(pa, "*Immunog*")
orde = setup_entities2(pa)
onto_plot2(orde, names(ol))
}
